requirements.txt
streamlit
requests
pandas
gspread
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
PyMuPDF
pdfplumber
Pillow
pytesseract
python-docx